[Vulkan][Runtime] Uniform buffer bugfix, minor cleanup (#7966)
* Bugfix, missing decoration on uniform buffer arguments.

Caused segfault when running on NVidia GPUs, with models that required
uniform buffer arguments for constants.

* Updated test_target_codegen_spirv.py to use @tvm.testing.requires_vulkan

Previously, these tests would show success if USE_VULKAN=OFF.  Now,
they correctly show that they are skipped instead.

* Minor cleanup on the vulkan runtime.

- Explicitly require int64 support at device creation time, since the
  TVM-generated shaders require it.

- Allocate an appropriate pool size for the buffer inputs, including
  both uniform and storage buffers.

* [Vulkan][Tests] Merged test_target_codegen_spirv.py into test_target_codegen_vulkan.py
